Project/cloudfunctions/delete_cloud_storage_image/requirements.txt
# Function dependencies, for example:
# package>=version
pymongo==4.3.3
dnspython
functions-framework==3.2.0
requests
google-cloud-storage